giccs: only accept a VirtualGlobbal.DirEnt in try_to_upload()
Previously it also accepted a pathlib.PurePath to avoid looking it up
in the GCSGlobber in the non-happy case.  This is more complicated than
it's worth.  This way we won't even attempt to upload a blob if one exists
already in our mind.

Also renamed and reordered the argumens of upload_file() and try_to_upload().