giccs: implement tab-completion of commands and files
On the top of the Readline library with a custom completer.
Since just by parsing the command line we cannot tell whether an argument
designates a local or remote file name, we offer matches for both.  Also
added a new INI setting to load the inputrc from.

This is all handled by the new Readline class.  In order to handle file names
with spaces and special characters correctly, we need to tokenize the input
while searching for matches, so the new class takes FTPClient.split() over
as Readline.split3().