giccs: introduce VirtualGlobber.DirEnt.Object
As the type of DirEnt.obj.  Now if the DirEnt is a file, it always has
an Object, either a MetaBlob or a handle through which it can be loaded.
Callers no longer need to check if the obj is None.

In future, if the Object is not a MetaBlob, it will be loaded automatically
on first access.

Object