giccs: add the --dircache flag to the ftp command
To store the directory hierarchy, so it doesn't have to be loaded from
the bucket.  Primarily intended for encrypted metadata, when all blobs
would have to be loaded just to retrieve the listing of a directory.

After loading the dircache we recreate the DirEnt hierarchy, but don't
load the blobs.  Currently they will be loaded when the blob property
of the DirEnt is accessed, so it works transparently.